+++ tests/Rector/Class_/AddHasFactoryToModelsRector/AddHasFactoryToModelsRectorTest.php
@@ -12,7 +12,7 @@ final class AddHasFactoryToModelsRectorTest extends AbstractRectorTestCase
 {
     public static function provideData(): Iterator
     {
-        return self::yieldFilesFromDirectory(__DIR__ . '/Fixture');
+        return self::yieldFilesFromDirectory(__DIR__ . '/Fixture/WithoutConfiguration');
     }
 
     /**
@@ -26,6 +26,6 @@ final class AddHasFactoryToModelsRectorTest extends AbstractRectorTestCase
 
     public function provideConfigFilePath(): string
     {
-        return __DIR__ . '/config/configured_rule.php';
+        return __DIR__ . '/config/configured_rule_without_configuration.php';
     }
 }
