+++ tests/Rector/Class_/AddHasFactoryToModelsRector/Fixture/Configured/fixture.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector;
+namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector\Fixture\Configured;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -12,7 +12,7 @@ class User extends Model
 -----
 <?php
 
-namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector;
+namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector\Fixture\Configured;
 
 use Illuminate\Database\Eloquent\Model;
