+++ tests/Rector/Class_/AddHasFactoryToModelsRector/Fixture/Configured/no_match.php.inc
@@ -1,0 +1,7 @@
+<?php
+
+namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector\Fixture\Configured;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Person extends Model {}
