+++ tests/Rector/Class_/AddHasFactoryToModelsRector/Fixture/Configured/skip_if_already_present.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRectorConfigured\Fixture;
+namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector\Fixture\Configured;
 
 use Illuminate\Database\Eloquent\Model;
