+++ tests/Rector/Class_/AddHasFactoryToModelsRector/Fixture/WithoutConfiguration/fixture.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRectorConfigured\Fixture;
+namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector\Fixture\WithoutConfiguration;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -12,7 +12,7 @@ class User extends Model
 -----
 <?php
 
-namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRectorConfigured\Fixture;
+namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector\Fixture\WithoutConfiguration;
 
 use Illuminate\Database\Eloquent\Model;
