+++ tests/Rector/Class_/AddHasFactoryToModelsRector/config/configured_rule_with_configuration.php
@@ -9,7 +9,7 @@ return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->import(__DIR__ . '/../../../../../config/config.php');
 
     $rectorConfig->ruleWithConfiguration(AddHasFactoryToModelsRector::class, [
-        'RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRectorConfigured\Fixture\User',
+        'RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRector\Fixture\Configured\User',
         'RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRectorConfigured\Fixture\SkipIfAlreadyPresent',
     ]);
 };
