+++ tests/Rector/Class_/AddHasFactoryToModelsRector/config/configured_rule_without_configuration.php
