--- tests/Rector/Class_/AddHasFactoryToModelsRectorConfigured/Fixture/no_match.php.inc
@@ -1,7 +1,0 @@
-<?php
-
-namespace RectorLaravel\Tests\Rector\Class_\AddHasFactoryToModelsRectorConfigured\Fixture;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Person extends Model {}
